+++ app/src/main/java/com/cokiri/coinkiri/presentation/createpost/CreatePostViewModel.kt
@@ -1,4 +1,4 @@
-package com.cokiri.coinkiri.presentation.post.community
+package com.cokiri.coinkiri.presentation.createpost
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -13,28 +13,24 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class CommunityWriteViewModel @Inject constructor(
+class CreatePostViewModel @Inject constructor(
     private val submitPostUseCase: SubmitPostUseCase
 ) : ViewModel() {
 
     // 게시글 제목을 관리하는 MutableStateFlow
     private val _title = MutableStateFlow("")
-    // 외부에서 읽기만 가능한 게시글 제목 StateFlow
     val title: StateFlow<String> = _title
 
     // 게시글 내용을 관리하는 MutableStateFlow
     private val _content = MutableStateFlow("")
-    // 외부에서 읽기만 가능한 게시글 내용 StateFlow
     val content: StateFlow<String> = _content
 
     // 로딩 상태를 관리하는 MutableStateFlow
     private val _isLoading = MutableStateFlow(false)
-    // 외부에서 읽기만 가능한 로딩 상태 StateFlow
     val isLoading: StateFlow<Boolean> = _isLoading
 
     // 에러 메시지를 관리하는 MutableStateFlow
     private val _errorMessage = MutableStateFlow<String?>(null)
-    // 외부에서 읽기만 가능한 에러 메시지 StateFlow
     val errorMessage: StateFlow<String?> = _errorMessage
 
     // 이미지 목록을 관리하는 MutableStateFlow
@@ -42,7 +38,6 @@ class CommunityWriteViewModel @Inject constructor(
 
     // 게시글 등록 결과를 관리하는 MutableStateFlow
     private val _submitResult = MutableStateFlow<Result<ApiResponse>?>(null)
-    // 외부에서 읽기만 가능한 게시글 등록 결과 StateFlow
     val submitResult: StateFlow<Result<ApiResponse>?> = _submitResult
 
     // 제목이 변경될 때 호출되는 함수
