+++ app/src/main/java/com/cokiri/coinkiri/util/Const.kt
@@ -7,13 +7,13 @@ const val HOME = "HOME"
 const val PRICE = "PRICE"
 const val PROFILE = "PROFILE"
 
+
 const val FOLLOW = "FOLLOW"
 
-
 const val COMMUNITY_DETAIL_SCREEN = "COMMUNITY_DETAIL_SCREEN"
-const val COMMUNITY_WRITE = "COMMUNITY_WRITE"
+
+const val CREATE_POST_SCREEN = "CREATE_POST_SCREEN"
 
 const val COIN_DETAIL = "COIN_DETAIL"
-
 
 const val ANALYSIS_WRITE_SCREEN = "ANALYSIS_WRITE_SCREEN"
